除了自然语言和单点登录

+++ src/main/java/org/example/backend/service/EventTimeService.java
@@ -22,7 +22,10 @@ public class EventTimeService {
     }
 
     public void delete(EventTime eventTime) {
-        eventTime.detach();
+        if(eventTime == null){
+            return;
+        }
+//        eventTime.detach();
         eventTimeRepository.delete(eventTime);
     }
 }
